requirements.txt
## Original requirements
faster-whisper>=1.0.3
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.11.0
rich>=13.7.0
python-dotenv>=1.0.0
hf_transfer>=0.1.9
huggingface_hub>=0.22.0

## Pipeline v2 additions
pyannote.audio>=3.1.0       # Speaker diarization (Piece 2)
pydantic>=2.0               # Data model validation across the entire pipeline
pyyaml>=6.0                 # config.yaml parsing
ollama>=0.2.0               # Ollama Python client for local LLM (Pieces 4 & 5)

## Testing
pytest>=8.0
pytest-mock>=3.14
